+++ tests/Helper/EscaperHelperTest.php
@@ -19,7 +19,7 @@ declare(strict_types=1);
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace Helper;
+namespace Tests\MacFJA\RediSearch\Helper;
 
 use function array_combine;
 use function array_map;